requirements.txt
Pillow
PyYAML
tqdm
